画面共通化処理を実装
